Level.java: move up to separate package
