+++ src/main/java/seedu/address/model/tuitionclass/TuitionClass.java
@@ -7,6 +7,7 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
+import seedu.address.model.level.Level;
 import seedu.address.model.tag.Tag;
 
 /**
